mercadopago/requirements.txt
fastapi==0.116.2
mercadopago==2.3.0
python-dotenv==1.1.1
uvicorn==0.35.0